requirements.txt
h5py==2.8.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
keras-tqdm==2.0.1
matplotlib==3.0.2
numpy==1.15.4
pandas==0.23.4
scipy==1.1.0
seaborn==0.9.0
tables==3.4.4
tensorboard==1.12.2
tensorflow==1.12.0
tqdm==4.28.1
Pillow

